Add UnsafeMiddleware
UnsafeMiddleware is a subclass of Middleware, designed to handle cases
where the middleware needs to access the store.  It's *unsafe* because
the middleware must not modify the store.  It's useful, because
middleware can need access to the store to instantiate
a ViewController. For example, when a routing action indicates that
a modal VC should be displayed, this allows middleware to handle that
action directly.